refactor(frontend): extract HistoryPanel from page.tsx (P1.e)
Moves the /history route UI into app/components/HistoryPanel.tsx.
page.tsx drops 105 lines net.

HistoryPanel owns its own document-id selection, jobs list, commits
list, and loading flag. Parent's \`history\` state and \`loadHistory\`
stay in page.tsx because they also drive the in-document history
drawer on the main review view — the two surfaces used to share that
state and could clobber each other when the user had different docs
selected in each view.

The panel picks a default document when libraryEntries populates
(async fetch post-mount) and tracks whether the user has made an
explicit selection so the auto-default does not overwrite their choice.

Verified: frontend 46/46 pass (stable across 5 full-suite reruns),
production build succeeds, /history route renders identically.
